ftest: add iterable dataset to tests
Features: DfuseFind
